# src/main/resources/swagger.json
{
    "openapi":"3.0.1",
    "info":{
       "title":"OpenAPI tetstetete",
       "version":"v0"
    },
    "servers":[
       {
          "url":"http://localhost:8080",
          "description":"Generated server url"
       }
    ],
    "paths":{
       "/v1/consulta-endereco":{
          "post":{
             "tags":[
                "endereco-controller"
             ],
             "operationId":"consultaEndereco",
             "requestBody":{
                "content":{
                   "application/json":{
                      "schema":{
                         "$ref":"#/components/schemas/EnderecoRequest"
                      }
                   }
                },
                "required":true
             },
             "responses":{
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/Endereco"
                         }
                      }
                   }
                }
             }
          }
       }
    },
    "components":{
       "schemas":{
          "EnderecoRequest":{
             "type":"object",
             "properties":{
                "cep":{
                   "type":"string"
                },
                "logradouro":{
                   "type":"string"
                },
                "complemento":{
                   "type":"string"
                },
                "bairro":{
                   "type":"string"
                },
                "localidade":{
                   "type":"string"
                },
                "uf":{
                   "type":"string"
                }
             }
          },
          "Endereco":{
             "type":"object",
             "properties":{
                "cep":{
                   "type":"string"
                },
                "rua":{
                   "type":"string"
                },
                "complemento":{
                   "type":"string"
                },
                "bairro":{
                   "type":"string"
                },
                "cidade":{
                   "type":"string"
                },
                "estado":{
                   "type":"string"
                },
                "frete":{
                   "type":"number",
                   "format":"double"
                }
             }
          }
       }
    }
 }